apiextensions: move openapi schema builder and merge code into package
